fixed useHookForm register not working when passes as a prop
