+++ buzzbridge-frontend/src/components/Forms/SignUpForm.tsx
@@ -1,16 +1,24 @@
-import { TextField, Typography, useMediaQuery } from "@mui/material";
+import {
+  FormControl,
+  FormHelperText,
+  IconButton,
+  InputAdornment,
+  InputLabel,
+  OutlinedInput,
+  TextField,
+  Typography,
+  useMediaQuery,
+} from "@mui/material";
 import React, { useState } from "react";
-import CustomPasswordInputField from "../Custom/CustomPasswordInputField";
 import CustomLoadingButton from "../Custom/CustomLoadingButton";
 import { useNavigate } from "react-router-dom";
 import useCustomAxios from "../../utils/helpers/customAxios";
 import { useAlert } from "../Providers/AlertProvider";
 import { ResetPassword } from "../../types/UserTypes";
-import {
-  ChangePasswordSchema,
-} from "../../utils/schema/userSchema";
+import { ChangePasswordSchema } from "../../utils/schema/userSchema";
 import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
+import { Visibility, VisibilityOff } from "@mui/icons-material";
 
 const SignUpForm = ({
   user,
@@ -30,6 +38,15 @@ const SignUpForm = ({
   const { showAlert } = useAlert();
   const [loading, setLoading] = useState(false);
   const [success, setSuccess] = useState<boolean | null>(null);
+
+  const [showPassword, setShowPassword] = useState(false);
+  const handleClickShowPassword = () => setShowPassword((show) => !show);
+
+  const handleMouseDownPassword = (
+    event: React.MouseEvent<HTMLButtonElement>
+  ) => {
+    event.preventDefault();
+  };
 
   const handleChange = async (e: any) => {
     setFormData((prev: any) => ({
@@ -92,24 +109,66 @@ const SignUpForm = ({
         BuzzBridge
       </Typography>
       {!forgetPassword && (
-        <TextField disabled label="Email" value={user.email || ""} />
+        <TextField disabled fullWidth label="Email" value={user.email || ""} />
       )}
-      <CustomPasswordInputField
-        config={register("newPassword", { required: true })}
-        name="newPassword"
-        label="New Password"
-        onBlur={handleChange}
-        error={Boolean(errors.newPassword?.message)}
-        helperText={errors.newPassword?.message}
-      />
-      <CustomPasswordInputField
-        config={register("confirmPassword", { required: true })}
-        name="confirmPassword"
-        label="Confirm Password"
-        onBlur={handleChange}
-        error={Boolean(errors.confirmPassword?.message)}
-        helperText={errors.confirmPassword?.message}
-      />
+      <FormControl sx={{ width: "300px" }} variant="outlined">
+        <InputLabel htmlFor="outlined-adornment-password">
+          New Password
+        </InputLabel>
+        <OutlinedInput
+          {...register("newPassword")}
+          fullWidth
+          name="newPassword"
+          label="New Password"
+          onBlur={handleChange}
+          error={Boolean(errors.newPassword?.message)}
+          type={showPassword ? "text" : "password"}
+          endAdornment={
+            <InputAdornment position="end">
+              <IconButton
+                aria-label="toggle password visibility"
+                onClick={handleClickShowPassword}
+                onMouseDown={handleMouseDownPassword}
+                edge="end"
+              >
+                {showPassword ? <VisibilityOff /> : <Visibility />}
+              </IconButton>
+            </InputAdornment>
+          }
+        />
+        <FormHelperText error id="username-error">
+          {errors.newPassword?.message}
+        </FormHelperText>
+      </FormControl>
+      <FormControl sx={{ width: "300px" }} variant="outlined">
+        <InputLabel htmlFor="outlined-adornment-password">
+          New Password
+        </InputLabel>
+        <OutlinedInput
+          fullWidth
+          {...register("confirmPassword")}
+          name="confirmPassword"
+          label="Confirm Password"
+          onBlur={handleChange}
+          error={Boolean(errors.confirmPassword?.message)}
+          type={showPassword ? "text" : "password"}
+          endAdornment={
+            <InputAdornment position="end">
+              <IconButton
+                aria-label="toggle password visibility"
+                onClick={handleClickShowPassword}
+                onMouseDown={handleMouseDownPassword}
+                edge="end"
+              >
+                {showPassword ? <VisibilityOff /> : <Visibility />}
+              </IconButton>
+            </InputAdornment>
+          }
+        />
+        <FormHelperText error id="username-error">
+          {errors.confirmPassword?.message}
+        </FormHelperText>
+      </FormControl>
       <CustomLoadingButton
         loading={displaySmallScreen ? false : loading}
         success={displaySmallScreen ? null : success}
